task5: statistic: fix bug with max value

+++ task5/05.homework/statistics.cpp
@@ -38,7 +38,7 @@ private:
 
 class Max : public IStatistics {
 public:
-	Max() : m_max{std::numeric_limits<double>::min()} {
+	Max() : m_max{ std::numeric_limits<int>::min()} {
 	}
 
 	void update(double next) override {
